Add documentation for MapField
