+++ crates/cgp-field/src/traits/mod.rs
@@ -2,10 +2,12 @@ mod from_fields;
 mod has_field;
 mod has_field_mut;
 mod has_fields;
+mod map_field;
 mod to_fields;
 
 pub use from_fields::*;
 pub use has_field::*;
 pub use has_field_mut::*;
 pub use has_fields::*;
+pub use map_field::*;
 pub use to_fields::*;
